+++ src/components/Register/RegisterPage.css
@@ -28,9 +28,9 @@
 
 /* desktop screen */
 @media (min-width: 600px) {
-    .register-form-page {   
-        /* width: 400px; */
-    }
+    /* .register-form-page {   
+        width: 400px;
+    } */
 
     .register-form-container {
         width: 400px;
@@ -42,9 +42,9 @@
 }
 
 @media (max-width: 600px) {
-    .register-form-page {    
+    /* .register-form-page {    
  
-    }
+    } */
 
     .register-form-container {
         width: 400px;
